requirements.txt
pandas==2.1.1
requests==2.31.0
pyarrow==13.0.0
boto3==1.28.36
python-dotenv==1.0.0
